pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "wearcapture"
version = "0.1.0"
description = "Cross-platform Wear OS long screenshot capture over ADB"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.24",
  "Pillow>=10.0",
]

[project.scripts]
wearcapture = "wearcapture.cli:main"

[tool.setuptools]
packages = ["wearcapture"]
